requirements.txt
streamlit==1.10.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.24.1
